auto detect local text editor nvim/vim/nano
This diff updates the shell profile scripts to look for what text
editors are installed, and picks nvim/vim/nano/vi in descending order of
preference.

Specifically $EDITOR will be set to one of nvim/vim/nano/vi depending on
what is installed on the system. And two new commands `v` and `vi` will
alias to the value set in $EDITOR.